requirements.txt
google-genai
pillow
python-dotenv
flask
flask-cors
gunicorn
exifread
psycopg2-binary
